requirements/test.txt
alembic==1.2.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
boto==2.49.0
boto3==1.10.5
botocore==1.13.5
btrdb==5.6
Cerberus==1.3.1
certifi==2019.9.11
Click==7.0
conduit==1.1
docutils==0.15.2
factory-boy==2.12.0
Faker==2.0.4
filedepot==0.7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Grinder-Webtest==0.1
grpcio==1.16.1
grpcio-tools==1.16.1
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.17.3
packaging==19.2
pandas==0.25.2
passlib==1.7.1
Pillow==6.2.1
pluggy==0.13.0
protobuf==3.10.0
psycopg2-binary==2.8.4
py==1.8.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.4.4
pytest==5.2.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
redis==3.3.11
s3transfer==0.2.1
scipy==1.3.1
sharedmem==0.3.7
six==1.12.0
soupsieve==1.9.5
SQLAlchemy==1.3.10
SQLAlchemy-Utils==0.34.2
SQLAlchemy-Wrapper==1.9.1
tabulate==0.8.5
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.25.6
waitress==1.3.1
wcwidth==0.1.7
WebOb==1.8.5
WebTest==2.0.33
Werkzeug==0.16.0
zipp==0.6.0
